backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==11.3.0
django-filter==25.1
drf-spectacular==0.27.1
python-dateutil==2.9.0
whitenoise==6.9.0